#992 reorganise plots
