+++ pybamm/plotting/quick_plot.py
@@ -44,114 +44,6 @@ def split_long_string(title, max_words=4):
         return first_line + "\n" + second_line
 
 
-def dynamic_plot(*args, **kwargs):
-    """
-    Creates a :class:`pybamm.QuickPlot` object (with arguments 'args' and keyword
-    arguments 'kwargs') and then calls :meth:`pybamm.QuickPlot.dynamic_plot`.
-    The key-word argument 'testing' is passed to the 'dynamic_plot' method, not the
-    `QuickPlot' class.
-
-    Returns
-    -------
-    plot : :class:`pybamm.QuickPlot`
-        The 'QuickPlot' object that was created
-    """
-    kwargs_for_class = {k: v for k, v in kwargs.items() if k != "testing"}
-    plot = pybamm.QuickPlot(*args, **kwargs_for_class)
-    plot.dynamic_plot(kwargs.get("testing", False))
-    return plot
-
-
-def plot(x, y, xlabel=None, ylabel=None, title=None, testing=False, **kwargs):
-    """
-    Generate a simple 1D plot. Calls `matplotlib.pyplot.plot` with keyword
-    arguments 'kwargs'.
-
-    Parameters
-    ----------
-    x : :class:`pybamm.Array`
-        The array to plot on the x axis
-    y : :class:`pybamm.Array`
-        The array to plot on the y axis
-    xlabel : str, optional
-        The label for the x axis
-    ylabel : str, optional
-        The label for the y axis
-    testing : bool, optional
-        Whether to actually make the plot (turned off for unit tests)
-
-    """
-    import matplotlib.pyplot as plt
-
-    if not isinstance(x, pybamm.Array):
-        raise TypeError("x must be 'pybamm.Array'")
-    if not isinstance(y, pybamm.Array):
-        raise TypeError("y must be 'pybamm.Array'")
-
-    plt.plot(x.entries, y.entries, **kwargs)
-    plt.ylim([ax_min(y.entries), ax_max(y.entries)])
-    plt.xlabel(xlabel)
-    plt.ylabel(ylabel)
-    plt.title(title)
-
-    if not testing:  # pragma: no cover
-        plt.show()
-
-    return
-
-
-def plot2D(x, y, z, xlabel=None, ylabel=None, title=None, testing=False, **kwargs):
-    """
-    Generate a simple 2D plot. Calls `matplotlib.pyplot.contourf` with keyword
-    arguments 'kwargs'.
-
-    Parameters
-    ----------
-    x : :class:`pybamm.Array`
-        The array to plot on the x axis
-    y : :class:`pybamm.Array`
-        The array to plot on the y axis
-    z : :class:`pybamm.Array`
-        The array to plot on the z axis
-    xlabel : str, optional
-        The label for the x axis
-    ylabel : str, optional
-        The label for the y axis
-    title : str, optional
-        The title for the plot
-    testing : bool, optional
-        Whether to actually make the plot (turned off for unit tests)
-
-    """
-    import matplotlib.pyplot as plt
-
-    if not isinstance(x, pybamm.Array):
-        raise TypeError("x must be 'pybamm.Array'")
-    if not isinstance(y, pybamm.Array):
-        raise TypeError("y must be 'pybamm.Array'")
-    if not isinstance(z, pybamm.Array):
-        raise TypeError("z must be 'pybamm.Array'")
-
-    plt.contourf(
-        x.entries,
-        y.entries,
-        z.entries,
-        vmin=ax_min(z.entries),
-        vmax=ax_max(z.entries),
-        cmap="coolwarm",
-        **kwargs
-    )
-    plt.xlabel(xlabel)
-    plt.ylabel(ylabel)
-    plt.title(title)
-    plt.colorbar()
-
-    if not testing:  # pragma: no cover
-        plt.show()
-
-    return
-
-
 class QuickPlot(object):
     """
     Generates a quick plot of a subset of key outputs of the model so that the model
